Add Employee/Voucher

+++ eventTom/eventTom.Api/Repositories/interfaces/IEventRepository.cs
@@ -4,7 +4,7 @@ namespace eventTom.Api.Repositories.interfaces
 {
     public interface IEventRepository
     {
-        Task<Event> GetByIdAsync(long id);
-        Task<IEnumerable<Event>> GetAllAsync();
+        Task<Event> GetByIdAsync(int id);                   // Event anhand der ID suchen
+        Task<IEnumerable<Event>> GetAllAsync();             // Alle Events abrufen
     }
 }
